Add CameraTransitionTests
