+++ Tests/MapboxMapsTests/Foundation/Camera/MapboxMapsCameraTests.swift
@@ -1,11 +1,5 @@
 import XCTest
-import MetalKit
-
-#if canImport(MapboxMaps)
 @testable import MapboxMaps
-#else
-@testable import MapboxMapsFoundation
-#endif
 
 internal class CameraManagerIntegrationTests: MapViewIntegrationTestCase {
 
@@ -82,23 +76,6 @@ internal class CameraManagerIntegrationTests: MapViewIntegrationTestCase {
         XCTAssertEqual(mapView.pitch, cameraManager.mapCameraOptions.maximumPitch, accuracy: 0.000001)
     }
 
-}
-
-class CameraManagerTests: XCTestCase {
-
-    var mapView: BaseMapView!
-    var cameraManager: CameraManager!
-    var mapInitOptions: MapInitOptions!
-
-    override func setUp() {
-        mapInitOptions = MapInitOptions(resourceOptions: ResourceOptions(accessToken: "pk.feedcafedeadbeefbadebede"))
-
-        mapView = BaseMapView(frame: CGRect(x: 0, y: 0, width: 100, height: 100),
-                              mapInitOptions: mapInitOptions,
-                              styleURI: nil)
-        cameraManager = CameraManager(for: mapView, with: MapCameraOptions())
-    }
-
     func testCameraForCoordinateArray() {
         // A 1:1 square
         let southwest = CLLocationCoordinate2DMake(0, 0)
@@ -126,73 +103,4 @@ class CameraManagerTests: XCTestCase {
         XCTAssertEqual(camera.pitch, 0)
     }
 
-    func testOptimizeBearingClockwise() {
-        let startBearing = 0.0
-        let endBearing = 90.0
-        let optimizedBearing = cameraManager.optimizeBearing(startBearing: startBearing, endBearing: endBearing)
-
-        XCTAssertEqual(optimizedBearing, 90.0)
-    }
-
-    func testOptimizeBearingCounterClockwise() {
-        let startBearing = 0.0
-        let endBearing = 270.0
-        let optimizedBearing = cameraManager.optimizeBearing(startBearing: startBearing, endBearing: endBearing)
-
-        // We should rotate counter clockwise which is shown by a negative angle
-        XCTAssertEqual(optimizedBearing, -90.0)
-    }
-
-    func testOptimizeBearingWhenBearingsAreTheSame() {
-        let startBearing = -90.0
-        let endBearing = 270.0
-        let optimizedBearing = cameraManager.optimizeBearing(startBearing: startBearing, endBearing: endBearing)
-
-        // -90 and 270 degrees is the same bearing so should just return original
-        XCTAssertEqual(optimizedBearing, -90)
-    }
-
-    func testOptimizeBearingWhenStartBearingIsNegative() {
-        var optimizedBearing: CLLocationDirection?
-
-        // Starting at -90 aka 270 should rotate clockwise to 20
-        optimizedBearing = cameraManager.optimizeBearing(startBearing: -90.0, endBearing: 20.0)
-        XCTAssertEqual(optimizedBearing, 20)
-
-        // Starting at -90 aka 270 should rotate clockwise to -270 aka 90
-        optimizedBearing = cameraManager.optimizeBearing(startBearing: -90.0, endBearing: -270)
-        XCTAssertEqual(optimizedBearing, 90)
-    }
-
-    func testOptimizeBearingHandlesNil() {
-        var optimizedBearing: CLLocationDirection?
-
-        // Test when no end bearing is provided
-        optimizedBearing = cameraManager.optimizeBearing(startBearing: 0.0, endBearing: nil)
-        XCTAssertNil(optimizedBearing)
-
-        // Test when no start bearing is provided
-        optimizedBearing = cameraManager.optimizeBearing(startBearing: nil, endBearing: 90)
-        XCTAssertNil(optimizedBearing)
-
-        // Test when no bearings are provided
-        optimizedBearing = cameraManager.optimizeBearing(startBearing: nil, endBearing: nil)
-        XCTAssertNil(optimizedBearing)
-    }
-
-    func testOptimizeBearingLargerThan360() {
-        var optimizedBearing: CLLocationDirection?
-
-        // 719 degrees is the same as 359 degrees. -1 should be returned because it is the shortest path from starting at 90
-        optimizedBearing = cameraManager.optimizeBearing(startBearing: 90.0, endBearing: 719)
-        XCTAssertEqual(optimizedBearing, -1.0)
-
-        // -195 should be returned because it is the shortest path from starting at 180
-        optimizedBearing = cameraManager.optimizeBearing(startBearing: 180, endBearing: -555)
-        XCTAssertEqual(optimizedBearing, 165)
-
-        // -160 should be returned because it is the shortest path from starting at 180
-        optimizedBearing = cameraManager.optimizeBearing(startBearing: 180, endBearing: -520)
-        XCTAssertEqual(optimizedBearing, 200)
-    }
 }
